Merged solution from Microsoft/MatterCenter branch

+++ tree/master/cloud/src/solution/Microsoft.Legal.MatterCenter.Web/Controllers/MatterController.cs
@@ -24,6 +24,7 @@ using Microsoft.Legal.MatterCenter.Repository;
 using Microsoft.Legal.MatterCenter.Models;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.Legal.MatterCenter.Web.Common;
+using System.Collections.Generic;
 #endregion
 namespace Microsoft.Legal.MatterCenter.Service
 {
@@ -35,7 +36,7 @@ namespace Microsoft.Legal.MatterCenter.Service
     public class MatterController : Controller
     {
         private ErrorSettings errorSettings;
-        private ISPOAuthorization spoAuthorization;
+        
         private IMatterCenterServiceFunctions matterCenterServiceFunctions;
         private MatterSettings matterSettings;
         private IMatterRepository matterRepositoy;
@@ -53,7 +54,7 @@ namespace Microsoft.Legal.MatterCenter.Service
         /// <param name="matterCenterServiceFunctions"></param>
         public MatterController(IOptionsMonitor<ErrorSettings> errorSettings,
             IOptionsMonitor<MatterSettings> matterSettings,
-            ISPOAuthorization spoAuthorization,
+            
             IMatterCenterServiceFunctions matterCenterServiceFunctions,
             IMatterRepository matterRepositoy,
             ICustomLogger customLogger, IOptionsMonitor<LogTables> logTables,
@@ -63,8 +64,7 @@ namespace Microsoft.Legal.MatterCenter.Service
             )
         {
             this.errorSettings = errorSettings.CurrentValue;
-            this.matterSettings = matterSettings.CurrentValue;
-            this.spoAuthorization = spoAuthorization;
+            this.matterSettings = matterSettings.CurrentValue;            
             this.matterCenterServiceFunctions = matterCenterServiceFunctions;
             this.matterRepositoy = matterRepositoy;
             this.customLogger = customLogger;
@@ -87,7 +87,7 @@ namespace Microsoft.Legal.MatterCenter.Service
             try
             {
                 //Get the authorization token from the Request header
-                spoAuthorization.AccessToken = HttpContext.Request.Headers["Authorization"];
+                
                 #region Error Checking                
                 ErrorResponse errorResponse = null;
                 
@@ -135,7 +135,7 @@ namespace Microsoft.Legal.MatterCenter.Service
             try
             {
                 //Get the authorization token from the Request header
-                spoAuthorization.AccessToken = HttpContext.Request.Headers["Authorization"];
+                
                 ErrorResponse errorResponse = null;
                 #region Error Checking                
                 if (searchRequestVM == null && searchRequestVM.Client == null && searchRequestVM.SearchObject == null)
@@ -179,7 +179,7 @@ namespace Microsoft.Legal.MatterCenter.Service
         {
             try
             {
-                spoAuthorization.AccessToken = HttpContext.Request.Headers["Authorization"];
+                
                 #region Error Checking                
                 ErrorResponse errorResponse = null;                
                 if (pinRequestMatterVM == null && pinRequestMatterVM.Client == null && pinRequestMatterVM.MatterData == null)
@@ -218,7 +218,7 @@ namespace Microsoft.Legal.MatterCenter.Service
         {
             try
             {
-                spoAuthorization.AccessToken = HttpContext.Request.Headers["Authorization"];
+                
                 #region Error Checking                
                 ErrorResponse errorResponse = null;
                 //if the token is not valid, immediately return no authorization error to the user
@@ -266,7 +266,7 @@ namespace Microsoft.Legal.MatterCenter.Service
         {
             try
             {
-                spoAuthorization.AccessToken = HttpContext.Request.Headers["Authorization"];
+                
 
                 #region Error Checking
                 ErrorResponse errorResponse = null;
@@ -305,7 +305,7 @@ namespace Microsoft.Legal.MatterCenter.Service
         {
             try
             {
-                spoAuthorization.AccessToken = HttpContext.Request.Headers["Authorization"];
+                
                 #region Error Checking                
                 ErrorResponse errorResponse = null;                
                 if (matterData == null && string.IsNullOrWhiteSpace(matterData.MatterUrl) && string.IsNullOrWhiteSpace(matterData.MatterName))
@@ -347,7 +347,7 @@ namespace Microsoft.Legal.MatterCenter.Service
         {
             try
             {
-                spoAuthorization.AccessToken = HttpContext.Request.Headers["Authorization"];
+                
 
                 #region Error Checking                
                 ErrorResponse errorResponse = null;
@@ -391,7 +391,7 @@ namespace Microsoft.Legal.MatterCenter.Service
         {
             try
             {
-                spoAuthorization.AccessToken = HttpContext.Request.Headers["Authorization"];
+                
                 #region Error Checking                
                 ErrorResponse errorResponse = null;
                 if (string.IsNullOrWhiteSpace(saveConfigurationsVM.SiteCollectionPath) && saveConfigurationsVM.MatterConfigurations == null)
@@ -427,7 +427,7 @@ namespace Microsoft.Legal.MatterCenter.Service
         {
             try
             {
-                spoAuthorization.AccessToken = HttpContext.Request.Headers["Authorization"];
+                
                 #region Error Checking                
                 ErrorResponse errorResponse = null;
                 if (string.IsNullOrWhiteSpace(siteCollectionPath))
@@ -462,7 +462,7 @@ namespace Microsoft.Legal.MatterCenter.Service
         [SwaggerResponse(HttpStatusCode.OK)]        
         public IActionResult CheckMatterExists([FromBody]MatterMetdataVM matterMetadataVM)
         {
-            spoAuthorization.AccessToken = HttpContext.Request.Headers["Authorization"];
+            
             GenericResponseVM genericResponse = ServiceUtility.GenericResponse(matterSettings.DeleteMatterCode, ServiceConstants.TRUE);
             var client = matterMetadataVM.Client;
             var matter = matterMetadataVM.Matter;
@@ -543,7 +543,7 @@ namespace Microsoft.Legal.MatterCenter.Service
         [SwaggerResponse(HttpStatusCode.OK)]        
         public IActionResult CheckSecurityGroupExists([FromBody]MatterInformationVM matterInformationVM)
         {
-            spoAuthorization.AccessToken = HttpContext.Request.Headers["Authorization"];
+            
             GenericResponseVM genericResponse = null;
             var client = matterInformationVM.Client;
             var matter = matterInformationVM.Matter;
@@ -610,7 +610,7 @@ namespace Microsoft.Legal.MatterCenter.Service
             var userid = matterInformation.UserIds;
             try
             {
-                spoAuthorization.AccessToken = HttpContext.Request.Headers["Authorization"];
+                
                 #region Error Checking                
                 ErrorResponse errorResponse = null;
                 if (matterInformation.Client == null && matterInformation.Matter == null && matterInformation.MatterDetails == null)
@@ -725,7 +725,7 @@ namespace Microsoft.Legal.MatterCenter.Service
         [SwaggerResponse(HttpStatusCode.OK)]
         public IActionResult Delete([FromBody] MatterVM matterVM)
         {
-            spoAuthorization.AccessToken = HttpContext.Request.Headers["Authorization"];
+            
             ErrorResponse errorResponse = null;
             if (null == matterVM && null == matterVM.Client && null == matterVM.Matter && string.IsNullOrWhiteSpace(matterVM.Client.Url) && string.IsNullOrWhiteSpace(matterVM.Matter.Name))
             {
@@ -753,7 +753,7 @@ namespace Microsoft.Legal.MatterCenter.Service
         {
             ErrorResponse errorResponse = null;
             GenericResponseVM genericResponseVM = null;
-            spoAuthorization.AccessToken = HttpContext.Request.Headers["Authorization"];
+            
             if (null == matterMetdataVM && null == matterMetdataVM.Client && null == matterMetdataVM.Matter && string.IsNullOrWhiteSpace(matterMetdataVM.Client.Url))
             {
                 errorResponse = new ErrorResponse()
@@ -804,52 +804,82 @@ namespace Microsoft.Legal.MatterCenter.Service
         [SwaggerResponse(HttpStatusCode.OK)]        
         public IActionResult AssignContentType([FromBody] MatterMetadata matterMetadata)
         {
-            spoAuthorization.AccessToken = HttpContext.Request.Headers["Authorization"];
-            ErrorResponse errorResponse = null;
-            if (null == matterMetadata && null == matterMetadata.Client && null == matterMetadata.Matter )
-            {                
-                errorResponse = new ErrorResponse()
+
+            GenericResponseVM genericResponse = null;
+            if (null == matterMetadata && null == matterMetadata.Client && null == matterMetadata.Matter && 
+                matterMetadata.ManagedColumnTerms==null)
+            {
+                genericResponse = new GenericResponseVM()
                 {
-                    Message = errorSettings.MessageNoInputs,
-                    ErrorCode = HttpStatusCode.BadRequest.ToString(),
-                    Description = "No input data is passed"
+                    Value = errorSettings.MessageNoInputs,
+                    Code = HttpStatusCode.BadRequest.ToString(),
+                    IsError = true
                 };
-                return matterCenterServiceFunctions.ServiceResponse(errorResponse, (int)HttpStatusCode.OK);               
+                
+                return matterCenterServiceFunctions.ServiceResponse(genericResponse, (int)HttpStatusCode.OK);               
             }
 
             // For each value in the list of Content Type Names
             // Add that content Type to the Library
             Matter matter = matterMetadata.Matter;
             Client client = matterMetadata.Client;
+
+            //ToDo: This values will come from the client. Once the UI is implemented, 
+            //This will be removed
+            var managedColumnTerms = new Dictionary<string, ManagedColumn>();
+            managedColumnTerms.Add("PracticeGroup", new ManagedColumn()
+            {
+                TermName = matterMetadata.PracticeGroupTerm.TermName,
+                Id = matterMetadata.PracticeGroupTerm.Id
+            });
+
+            managedColumnTerms.Add("AreaOfLaw", new ManagedColumn()
+            {
+                TermName = matterMetadata.AreaTerm.TermName,
+                Id = matterMetadata.AreaTerm.Id
+            });
+
+            managedColumnTerms.Add("SubareaOfLaw", new ManagedColumn()
+            {
+                TermName = matterMetadata.SubareaTerm.TermName,
+                Id = matterMetadata.SubareaTerm.Id
+            });
+
+            matterMetadata.ManagedColumnTerms = managedColumnTerms;
+
             var matterInformationVM = new MatterInformationVM()
             {
                 Client = client,
                 Matter = matter,
-
             };
             try
             {                
-                GenericResponseVM genericResponse = validationFunctions.IsMatterValid(matterInformationVM, int.Parse(ServiceConstants.ProvisionMatterAssignContentType, CultureInfo.InvariantCulture), null);
+                genericResponse = validationFunctions.IsMatterValid(matterInformationVM, int.Parse(ServiceConstants.ProvisionMatterAssignContentType, 
+                    CultureInfo.InvariantCulture), null);
                 if (genericResponse != null)
                 { 
                     matterProvision.DeleteMatter(matterInformationVM as MatterVM);
-                    errorResponse = new ErrorResponse()
+                    
+                    genericResponse = new GenericResponseVM()
                     {
-                        Message = genericResponse.Value,
-                        ErrorCode = genericResponse.Code.ToString()                       
+                        Value = genericResponse.Value,
+                        Code = genericResponse.Code.ToString(),
+                        IsError = true
                     };
-                    return matterCenterServiceFunctions.ServiceResponse(errorResponse, (int)HttpStatusCode.OK);
+
+                    return matterCenterServiceFunctions.ServiceResponse(genericResponse, (int)HttpStatusCode.OK);
                 }
                 genericResponse = matterProvision.AssignContentType(matterMetadata);
                 if (genericResponse != null && genericResponse.IsError==true)
                 {
                     matterProvision.DeleteMatter(matterInformationVM as MatterVM);
-                    errorResponse = new ErrorResponse()
+                    genericResponse = new GenericResponseVM()
                     {
-                        Message = genericResponse.Value,
-                        ErrorCode = genericResponse.Code.ToString()
+                        Value = genericResponse.Value,
+                        Code = genericResponse.Code.ToString(),
+                        IsError = true
                     };
-                    return matterCenterServiceFunctions.ServiceResponse(errorResponse, (int)HttpStatusCode.OK);
+                    return matterCenterServiceFunctions.ServiceResponse(genericResponse, (int)HttpStatusCode.OK);
                 }
                 return matterCenterServiceFunctions.ServiceResponse(genericResponse, (int)HttpStatusCode.OK);
             }
@@ -875,7 +905,7 @@ namespace Microsoft.Legal.MatterCenter.Service
             var matter = matterMetadataVM.Matter;
             try
             {
-                spoAuthorization.AccessToken = HttpContext.Request.Headers["Authorization"];
+                
                 var matterConfigurations = matterMetadataVM.MatterConfigurations;
                 ErrorResponse errorResponse = null;
                 if (null == client && null == matter && null == client.Url && null == matterConfigurations)
@@ -924,7 +954,7 @@ namespace Microsoft.Legal.MatterCenter.Service
         [SwaggerResponse(HttpStatusCode.OK)]        
         public IActionResult CreateLandingPage([FromBody] MatterMetdataVM matterMetdataVM)
         {
-            spoAuthorization.AccessToken = HttpContext.Request.Headers["Authorization"];
+            
             ErrorResponse errorResponse = null;
             GenericResponseVM genericResponseVM = null;
             //No valid input
@@ -985,21 +1015,20 @@ namespace Microsoft.Legal.MatterCenter.Service
 
             try
             {
-                spoAuthorization.AccessToken = HttpContext.Request.Headers["Authorization"];
-
                 #region Error Checking                
-                ErrorResponse errorResponse = null;
+                GenericResponseVM genericResponse = null;
                 if (matterMetdata.Client == null && matterMetdata.Matter == null &&
-                    matterMetdata.MatterDetails == null && matterMetdata.MatterProvisionFlags == null)
+                    matterMetdata.MatterDetails == null && matterMetdata.MatterProvisionFlags == null && 
+                    matterMetdata.MatterDetails.ManagedColumnTerms==null)
                 {
                     matterProvision.DeleteMatter(matterMetdata as MatterVM);
-                    errorResponse = new ErrorResponse()
+                    genericResponse = new GenericResponseVM()
                     {
-                        Message = errorSettings.MessageNoInputs,
-                        ErrorCode = HttpStatusCode.BadRequest.ToString(),
-                        Description = "No input data is passed"
+                        Value = errorSettings.MessageNoInputs,
+                        Code = HttpStatusCode.BadRequest.ToString(),
+                        IsError = true
                     };
-                    return matterCenterServiceFunctions.ServiceResponse(errorResponse, (int)HttpStatusCode.BadRequest);
+                    return matterCenterServiceFunctions.ServiceResponse(genericResponse, (int)HttpStatusCode.OK);
                 }
                 #endregion
 
@@ -1010,23 +1039,47 @@ namespace Microsoft.Legal.MatterCenter.Service
                     Matter = matterMetdata.Matter,
                     MatterDetails = matterMetdata.MatterDetails
                 };
-                GenericResponseVM genericResponse = validationFunctions.IsMatterValid(matterInfo,
+                genericResponse = validationFunctions.IsMatterValid(matterInfo,
                     int.Parse(ServiceConstants.ProvisionMatterUpdateMetadataForList),
                     matterMetdata.MatterConfigurations);
                 if (genericResponse != null)
                 {
                     matterProvision.DeleteMatter(matterMetdata as MatterVM);
-                    errorResponse = new ErrorResponse()
+                    genericResponse = new GenericResponseVM()
                     {
-                        Message = genericResponse.Value,
-                        ErrorCode = genericResponse.Code,
-                    };
-                    return matterCenterServiceFunctions.ServiceResponse(errorResponse, (int)HttpStatusCode.BadRequest);
+                        Value = genericResponse.Value,
+                        Code = genericResponse.Code,
+                        IsError = true
+                    };                    
+                    return matterCenterServiceFunctions.ServiceResponse(genericResponse, (int)HttpStatusCode.OK);
                 }
                 #endregion   
 
                 try
                 {
+
+                    //ToDo: This values will come from the client. Once the UI is implemented, 
+                    //This will be removed
+                    var managedColumnTerms = new Dictionary<string, ManagedColumn>();
+                    managedColumnTerms.Add("PracticeGroup", new ManagedColumn()
+                    {
+                        TermName = matterMetdata.MatterDetails.PracticeGroup
+                    });
+
+                    managedColumnTerms.Add("AreaOfLaw", new ManagedColumn()
+                    {
+                        TermName = matterMetdata.MatterDetails.AreaOfLaw,
+
+                    });
+
+                    managedColumnTerms.Add("SubareaOfLaw", new ManagedColumn()
+                    {
+                        TermName = matterMetdata.MatterDetails.SubareaOfLaw,
+
+                    });
+
+                    matterMetdata.MatterDetails.ManagedColumnTerms = managedColumnTerms;
+
                     genericResponse = matterProvision.UpdateMatterMetadata(matterMetdata);
                     if (genericResponse == null)
                     {
@@ -1067,7 +1120,7 @@ namespace Microsoft.Legal.MatterCenter.Service
             var client = matterInformation.Client;            
             try
             {
-                spoAuthorization.AccessToken = HttpContext.Request.Headers["Authorization"];
+                
                 
                 ErrorResponse errorResponse = null;
                 if (matterInformation == null && matterInformation.Client==null)
